style(MobileHeader): move component file to MobileHeader folder
