+++ src/components/MobilePreview/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import MobileFrame from './MobileFrame';
-import MobileHeader from './MobileHeader';
+import MobileHeader from '../../components/MobileHeader';
 
 const MobilePreview = () => (
   <MobileFrame>
